backend/requirements.txt
fastapi>=0.128.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pyjwt>=2.10.1
bcrypt==4.2.0
psycopg2-binary==2.9.9
email-validator==2.2.0
pydantic[email]>=2.12.5
pytest==8.3.4
httpx==0.28.1
openai>=2.14.0
openai-agents>=0.6.5
openai-agents-mcp>=0.0.8
mcp>=1.25.0
mcp-agent>=0.2.6